refactor presenter
