+++ package.json
@@ -33,7 +33,7 @@
     "eslint-plugin-promise": "^6.0.0",
     "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-react-hooks": "^4.5.0",
-    "husky": ">=6",
+    "husky": "^7.0.4",
     "jest": "^27.5.1",
     "lint-staged": ">=10",
     "prettier": "^2.6.2",
