--- src/view/components/features/todo/TodoPresenter.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import {TodoList} from './TodoList';
-import {useFetchTodos} from '../../../../presenter/hooks/todo/useFetchTodos';
-
-export const TodoPresenter = () => {
-  const {todos} = useFetchTodos();
-
-  return <TodoList todos={todos} />;
-};
